+++ Runtime/Scripts/Schema/Texture.cs
@@ -13,6 +13,8 @@
 // limitations under the License.
 //
 
+using System;
+
 namespace GLTFast.Schema {
 
     [System.Serializable]
@@ -58,6 +60,35 @@ namespace GLTFast.Schema {
                 extensions.GltfSerialize(writer);
             }
             writer.Close();
+        }
+
+        public override bool Equals(object obj) {
+            //Check for null and compare run-time types.
+            if (obj == null || ! GetType().Equals(obj.GetType())) {
+                return false;
+            }
+            return Equals((Texture)obj);
+        }
+        
+        bool Equals(Texture other) {
+            return source == other.source
+                && sampler == other.sampler
+                && (
+                    extensions == null && other.extensions == null
+                    || (extensions != null && extensions.Equals(other.extensions))
+                );
+        }
+
+        public override int GetHashCode() {
+#if NET_STANDARD
+            return HashCode.Combine(source, sampler, extensions.GetHashCode());
+#else
+                var hash = 17;
+                hash = hash * 31 + source;
+                hash = hash * 31 + sampler;
+                hash = hash * 31 + extensions.GetHashCode();
+                return hash;
+#endif
         }
     }
 }