+++ ChangeLog.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Misformated XML documentation comments
 - Correct render pipeline detection in case of quality settings override
 - (Documentation) Many minor fixes like XML doc linter errors/warnings
+- (Export) Removed redundant texture entries in glTF schema
 
 ## [4.5.0] - 2022-01-24
 ### Added